Fixed Operation Bugs

+++ rentermatch/src/main/java/com/humanlearning/rentermatch/utility/JWTUtility.java
@@ -15,7 +15,7 @@ import org.springframework.stereotype.Component;
 @Component
 public class JWTUtility implements Serializable {
 
-  public static final long JWT_TOKEN_VALIDITY = 5 * 60 * 60;
+  public static final long JWT_TOKEN_VALIDITY = (long)5 * 60 * 60;
   private static final long serialVersionUID = 234234523523L;
   @Value("${jwt.secret}")
   private String secretKey;
